reorganize types
