+++ packages/js/product-editor/src/components/button-with-dropdown-menu/index.tsx
@@ -4,50 +4,10 @@
 import { createElement } from '@wordpress/element';
 import { chevronDown } from '@wordpress/icons';
 import { Button, DropdownMenu, Flex, FlexItem } from '@wordpress/components';
-
-import type {
-	// @ts-expect-error no exported member.
-	DropdownOption,
-} from '@wordpress/components';
-
-type PositionYAxis = 'top' | 'middle' | 'bottom';
-type PositionXAxis = 'left' | 'center' | 'right';
-type PositionCorner = 'top' | 'right' | 'bottom' | 'left';
-
-type PopoverPlacement =
-	| 'left'
-	| 'right'
-	| 'bottom'
-	| 'top'
-	| 'left-end'
-	| 'left-start'
-	| 'right-end'
-	| 'right-start'
-	| 'bottom-end'
-	| 'bottom-start'
-	| 'top-end'
-	| 'top-start'; // @todo: pick from core
-
-type popoverPosition =
-	| `${ PositionYAxis }`
-	| `${ PositionYAxis } ${ PositionXAxis }`
-	| `${ PositionYAxis } ${ PositionXAxis } ${ PositionCorner }`;
-
-type popoverProps = {
-	placement?: PopoverPlacement;
-	position?: popoverPosition;
-	offset?: number;
-};
-
-export interface ButtonWithDropdownMenuProps {
-	label: string;
-	variant?: Button.ButtonProps[ 'variant' ];
-	defaultOpen?: boolean;
-	controls?: DropdownOption[];
-
-	popoverProps?: popoverProps;
-	onButtonClick?: () => void;
-}
+/**
+ * Internal dependencies
+ */
+import type { ButtonWithDropdownMenuProps } from './types';
 
 export const ButtonWithDropdownMenu: React.FC<
 	ButtonWithDropdownMenuProps
